requirements.txt
beautifulsoup4
selenium
mysql-connector-python
pandas
prettytable